fix import issue, move run chain functions
