+++ agixt/Chain.py
@@ -12,7 +12,8 @@ from DB import (
     User,
 )
 from Globals import getenv, DEFAULT_USER
-from ApiClient import Prompts, Conversations
+from Prompts import Prompts
+from Conversations import Conversations
 from Extensions import Extensions
 import logging
 
@@ -710,188 +711,6 @@ class Chain:
         self.session.add(chain_run)
         self.session.commit()
         return chain_run.id
-
-    async def run_chain_step(
-        self,
-        chain_run_id=None,
-        step: dict = {},
-        chain_name="",
-        user_input="",
-        agent_override="",
-        chain_args={},
-        conversation_name="",
-    ):
-        if not chain_run_id:
-            chain_run_id = await self.get_chain_run_id(chain_name=chain_name)
-        if step:
-            if "prompt_type" in step:
-                c = None
-                if conversation_name != "":
-                    c = Conversations(
-                        conversation_name=conversation_name,
-                        user=self.user,
-                    )
-                if agent_override != "":
-                    agent_name = agent_override
-                else:
-                    agent_name = step["agent_name"]
-                prompt_type = step["prompt_type"]
-                step_number = step["step"]
-                if "prompt_name" in step["prompt"]:
-                    prompt_name = step["prompt"]["prompt_name"]
-                else:
-                    prompt_name = ""
-                args = self.get_step_content(
-                    chain_run_id=chain_run_id,
-                    chain_name=chain_name,
-                    prompt_content=step["prompt"],
-                    user_input=user_input,
-                    agent_name=step["agent_name"],
-                )
-                if chain_args != {}:
-                    for arg, value in chain_args.items():
-                        args[arg] = value
-                if "chain_name" in args:
-                    args["chain"] = args["chain_name"]
-                if "chain" not in args:
-                    args["chain"] = chain_name
-                if "conversation_name" not in args:
-                    args["conversation_name"] = f"Chain Execution History: {chain_name}"
-                if "conversation" in args:
-                    args["conversation_name"] = args["conversation"]
-                if prompt_type == "Command":
-                    if conversation_name != "":
-                        c.log_interaction(
-                            role=agent_name,
-                            message=f"[ACTIVITY_START] `{agent_name}` is running command `{step['prompt']['command_name']}`... [ACTIVITY_END]",
-                        )
-                    result = self.ApiClient.execute_command(
-                        agent_name=agent_name,
-                        command_name=step["prompt"]["command_name"],
-                        command_args=args,
-                        conversation_name=args["conversation_name"],
-                    )
-                elif prompt_type == "Prompt":
-                    if conversation_name != "":
-                        c.log_interaction(
-                            role=agent_name,
-                            message=f"[ACTIVITY_START] `{agent_name}` is running prompt `{prompt_name}`... [ACTIVITY_END]",
-                        )
-                    result = self.ApiClient.prompt_agent(
-                        agent_name=agent_name,
-                        prompt_name=prompt_name,
-                        prompt_args={
-                            "user_input": user_input,
-                            **args,
-                        },
-                    )
-                elif prompt_type == "Chain":
-                    if conversation_name != "":
-                        c.log_interaction(
-                            role=agent_name,
-                            message=f"[ACTIVITY_START] `{agent_name}` is running chain `{args['chain']}`... [ACTIVITY_END]",
-                        )
-                    result = self.ApiClient.run_chain(
-                        chain_name=args["chain"],
-                        user_input=args["input"],
-                        agent_name=agent_name,
-                        all_responses=(
-                            args["all_responses"] if "all_responses" in args else False
-                        ),
-                        from_step=args["from_step"] if "from_step" in args else 1,
-                        chain_args=(
-                            args["chain_args"]
-                            if "chain_args" in args
-                            else {"conversation_name": args["conversation_name"]}
-                        ),
-                    )
-        if result:
-            if isinstance(result, dict) and "response" in result:
-                result = result["response"]
-            if result == "Unable to retrieve data.":
-                result = None
-            if not isinstance(result, str):
-                result = str(result)
-            await self.update_step_response(
-                chain_run_id=chain_run_id,
-                chain_name=chain_name,
-                step_number=step_number,
-                response=result,
-            )
-            return result
-        else:
-            return None
-
-    async def run_chain(
-        self,
-        chain_name,
-        chain_run_id=None,
-        user_input=None,
-        agent_override="",
-        from_step=1,
-        chain_args={},
-        log_user_input=False,
-        conversation_name="",
-    ):
-        chain_data = self.get_chain(chain_name=chain_name)
-        # Create a new ChainRun object to store the chain run data.
-        if not chain_run_id:
-            chain_run = ChainRun(
-                chain_id=chain_data["id"],
-                user_id=self.user,
-            )
-            self.session.add(chain_run)
-            self.session.commit()
-            chain_run_id = chain_run.id
-        if chain_data == {}:
-            return f"Chain `{chain_name}` not found."
-        c = Conversations(
-            conversation_name=conversation_name,
-            user=self.user,
-        )
-        if log_user_input:
-            c.log_interaction(
-                role="USER",
-                message=user_input,
-            )
-        agent_name = agent_override if agent_override != "" else "AGiXT"
-        if conversation_name != "":
-            c.log_interaction(
-                role=agent_name,
-                message=f"[ACTIVITY_START] Running chain `{chain_name}`... [ACTIVITY_END]",
-            )
-        last_response = ""
-        for step_data in chain_data["steps"]:
-            if int(step_data["step"]) >= int(from_step):
-                if "prompt" in step_data and "step" in step_data:
-                    step = {}
-                    step["agent_name"] = (
-                        agent_override
-                        if agent_override != ""
-                        else step_data["agent_name"]
-                    )
-                    step["prompt_type"] = step_data["prompt_type"]
-                    step["prompt"] = step_data["prompt"]
-                    step["step"] = step_data["step"]
-                    step_response = await self.run_chain_step(
-                        chain_run_id=chain_run_id,
-                        step=step,
-                        chain_name=chain_name,
-                        user_input=user_input,
-                        agent_override=agent_override,
-                        chain_args=chain_args,
-                        conversation_name=conversation_name,
-                    )
-                    if step_response == None:
-                        return f"Chain failed to complete, it failed on step {step_data['step']}. You can resume by starting the chain from the step that failed with chain ID {chain_run_id}."
-                    last_response = step_response
-                    logging.info(f"Step {step_data['step']} response: {step_response}")
-        if conversation_name != "":
-            c.log_interaction(
-                role=agent_name,
-                message=last_response,
-            )
-        return last_response
 
     def get_chain_args(self, chain_name):
         skip_args = [
