+++ agixt/Models.py
@@ -117,6 +117,7 @@ class RunChainStep(BaseModel):
     prompt: str
     agent_override: Optional[str] = ""
     chain_args: Optional[dict] = {}
+    chain_run_id: Optional[str] = ""
 
 
 class StepInfo(BaseModel):
